@@ -33,56 +33,64 @@
 		</developer>
 	</developers>
 	<scm>
-		<connection>scm:git:git@github.com:ggrandes/${project.artifactId}.git</connection>
-		<developerConnection>scm:git:git@github.com:ggrandes/${project.artifactId}.git</developerConnection>
-		<url>git@github.com:ggrandes/${project.artifactId}.git</url>
+		<connection>scm:git:git@github.com:ggrandes/concurrentlimit-servlet-filter.git</connection>
+		<developerConnection>scm:git:git@github.com:ggrandes/concurrentlimit-servlet-filter.git</developerConnection>
+		<url>git@github.com:ggrandes/concurrentlimit-servlet-filter.git</url>
 	</scm>
+
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>1.8</java.version>
+	</properties>
 
 	<build>
 		<plugins>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>2.3.2</version>
+				<version>3.10.1</version>
 				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
+					<source>${java.version}</source>
+					<target>${java.version}</target>
 				</configuration>
 			</plugin>
 			<!-- Package Classes to JAR to upload to repo -->
 			<plugin>
 				<artifactId>maven-jar-plugin</artifactId>
-				<version>2.3.1</version>
+				<version>3.2.2</version>
 				<configuration>
 					<finalName>${project.artifactId}-${project.version}</finalName>
 					<forceCreation>true</forceCreation>
 					<archive>
+						<manifest>
+							<addDefaultImplementationEntries>true</addDefaultImplementationEntries>
+						</manifest>
 						<manifestEntries>
 							<Built-By>maven</Built-By>
+							<Implementation-Vendor>${project.groupId}</Implementation-Vendor>
+							<Implementation-Title>${project.groupId}:${project.artifactId}</Implementation-Title>
+							<Implementation-Version>${project.version}</Implementation-Version>
 						</manifestEntries>
 					</archive>
-					<excludes>
-						<exclude>**/pom.properties</exclude>
-						<exclude>**/log4j.properties</exclude>
-						<exclude>**/sandbox/*</exclude>
-					</excludes>
 				</configuration>
 			</plugin>
 			<!-- Package source to JAR to upload to repo -->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-source-plugin</artifactId>
-				<version>2.2.1</version>
+				<version>3.2.1</version>
 				<executions>
 					<execution>
 						<goals>
 							<goal>jar</goal>
 						</goals>
 						<configuration>
+							<archive>
+								<manifestEntries>
+									<Built-By>maven</Built-By>
+								</manifestEntries>
+							</archive>
 							<excludeResources>true</excludeResources>
-							<excludes>
-								<exclude>**/sandbox/*</exclude>
-							</excludes>
 						</configuration>
 					</execution>
 				</executions>
